requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-weasyprint==1.0.2
gunicorn==20.0.4
html5lib==1.1
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
Pyphen==0.10.0
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52.1
webencodings==0.5.1
whitenoise==5.2.0
